Move getPendingTransactions to transactions store
