+++ services/core/methods/controllers/transactions.js
@@ -162,10 +162,10 @@ const getPendingTransactions = async (params) => {
 	const result = await CoreService.getPendingTransactions(params);
 	if (result) {
 		const meta = {
+			total: result.meta.count,
 			count: result.data.length,
 			limit: result.meta.limit || result.data.length,
 			offset: result.meta.offset || 0,
-			total: result.data.length,
 		};
 		return {
 				data: result.data,
